move css styles of settings-tab to its own file
